チャンネルのメッセージを取得するAPIを追加
